Added Base screen to UI module and used Material themeing throughout

+++ modules/feature/btconnection/src/main/java/com/esightcorp/mobile/app/btconnection/NoDevicesConnected.kt
@@ -67,8 +67,8 @@ internal fun NoDeviceConnectedScreen(
     onSettingsButtonPressed: () -> Unit,
     onFeedbackButtonPressed: () -> Unit,
     onConnectToDeviceButtonPressed: () -> Unit,
-    onTermsAndConditionsPressed: () -> Unit,
-    onPrivacyPolicyPressed: () -> Unit,
+    onTermsAndConditionsPressed: (Int) -> Unit,
+    onPrivacyPolicyPressed: (Int) -> Unit,
     btUiState: BluetoothUiState,
     navController: NavController,
     modifier: Modifier = Modifier
